Create LICENSE.md
We Have adopted the MIT License format which grants a variety of permissions to anyone who might want to use any of our work present in the repository. This allows us to maintain an open source format, promotes the use of our work, and encourages anyone interested in software and website development to implement or review our work with out much scrutiny. Lastly, while this format is incredibly liberal with it's use, it does not provide any warranty from us, in addition from absolving our software from legal issues that might arise or be caused by our work.